Add files via upload
Included new stats in spreadsheet to reduce need to traverse spreadsheet and computation leading to reduced runtime when commands are entered
Standardized formulas for all stats to be rounded to 2 decimals and be based off of players' total rounds in events rather than avg of stats at end of games. This is to prevent skewing of results as rounds are not equal from game to game which could cause abnormal stats(think HS%).   
Added new columns in spreadsheet for MVP, EVP and Placements. These are used to give event info with a new function - eventStats()
eventStats() - allows users to pull details of an event such as mvp, evps, winner, etc.

Once again some of this is framework and still needs to be tested/implement fully.
Added comments at top to signify future plans/idea