Add files via upload
Have input added now to ask user if they want to see a specific person's stats. This will loop after input unless broken, user is given instructions on how to break program out.
Input is parsed to see if user is in spreadsheet, if not a list of available users is given. Tested that it should work as intended.